:hammer: adicionado no final do componente o metodo de usar
